requirements.txt.txt
twitchio==2.6.0
python-dotenv